.ssh/config: added the proxy helper for Google stuff
N.B., we really want to develop some sort of tool for automatically
creating the ~/.ssh/config file from various pieces, so that we can
have different ones for different hosts. SSH's config format itself
doesn't have any way of doing that sort of thing, which is why we
need to make out own tool for it.